+++ ts-lgp-30-html5/src/orders/b.ts
@@ -1,8 +1,9 @@
-import type { BitArray, SectorNumber, IState, TrackNumber } from "../types";
+import type { BitArray, IState } from "../types";
+import type { TrackNumber, SectorNumber } from "../types/numbers";
 import { IOrder } from "./order";
 
 //Bring: regs.a = mem[track:sector]
-export class OrderB implements IOrder{
+export class OrderB implements IOrder {
     name = "b"
     orderNumber = [0, 0, 0, 1] as BitArray