+++ ts-lgp-30-html5/src/orders/h.ts
@@ -1,8 +1,9 @@
-import type { BitArray, SectorNumber, IState, TrackNumber } from "../types";
+import type { BitArray, IState } from "../types";
+import type { TrackNumber, SectorNumber } from "../types/numbers";
 import { IOrder } from "./order";
 
 //Hold and Store: mem[track:sector] = regs.a
-export class OrderH implements IOrder{
+export class OrderH implements IOrder {
     name = "h"
     orderNumber = [1, 1, 0, 0] as BitArray