+++ ts-lgp-30-html5/src/regs/c.ts
@@ -1,5 +1,7 @@
-import { binToDec, halfToHex, decToBin, hexToDec } from "../util"
+import { halfToHex, hexToDec } from "../util"
 import { Register } from "./register"
+import type { BitArray } from "../types"
+import { asSector, asTrack, toBits } from "../types/numbers"
 
 export class RegisterC extends Register {
     constructor() {
@@ -8,19 +10,26 @@ export class RegisterC extends Register {
 
     inc() {
         //Increment, respectinv TTSS addressing
-        let track = binToDec(this.data.slice(0, 6))
-        let sector = binToDec(this.data.slice(6, 13))
-
+        let track = asTrack(this.getTrack())
+        let sector = asSector(this.getSector())
         sector++
-
         if (sector > 63) {
-            sector = 0
+            sector = asSector(0)
             track++
         }
-
         // TODO what happens if track > 63?
+        const trackBits = toBits(track)
+        const sectorBits = toBits(sector)
+        const newBits = trackBits.concat(sectorBits)
+        this.data = newBits
+    }
 
-        this.data = decToBin(track, 6).concat(decToBin(sector, 6))
+    getTrack(): BitArray {
+        return this.get(0, 6)
+    }
+
+    getSector(): BitArray {
+        return this.get(6, 13)
     }
 
     getHexTrack() {
@@ -32,9 +41,12 @@ export class RegisterC extends Register {
     }
 
     toDec() {
-        const track = hexToDec(this.getHexTrack())
-        const sector = hexToDec(this.getHexSector())
+        return asTrack(this.getTrack()) * 64 + asSector(this.getSector())
+    }
 
-        return parseInt(track.toString() + (sector < 10 ? "0" : "") + sector.toString())
+    toTTSS(): string {
+        const trackStr = asTrack(this.getTrack()).toString().padEnd(2, '0')
+        const sectorStr = asSector(this.getSector()).toString().padEnd(2, '0')
+        return trackStr + sectorStr
     }
 }