+++ ts-lgp-30-html5/src/util.ts
@@ -5,18 +5,25 @@ import type { BitArray } from "./types";
 /**
  * Number encoding/decoding stuff
  */
-export function binToDec(b) {
+
+/**
+ * Convert an array of bits to a decimal number
+ * @param b The array of bits
+ * @returns The bits interpreted as a decimal number
+ */
+export function binToDec(b: BitArray): number {
     return parseInt(b.join(""), 2);
 }
 
 export function decToBin(d, n) {
+    // console.debug('decToBin d', d, 'n', n);
     const bits = d.toString(2).split("").map(Number)
-
     if (n == undefined) {
         return bits
     }
-
-    return Array(n - bits.length).fill(0).concat(bits)
+    const arrayLength = n - bits.length
+    if (arrayLength < 0) { throw new Error(`arrayLength ${arrayLength} < 0`) }
+    return Array(arrayLength).fill(0).concat(bits)
 }
 
 export function hexToDec(h) {
@@ -100,25 +107,35 @@ export function isHex(c) {
     return c == 0 || Number(c) || ["f", "g", "j", "k", "q", "w", "l"].includes(c)
 }
 
-//Encode literal number
-export function packNum(d) {
+/**
+ * Encode literal number (used for the registers)
+ * @param d the signed number
+ * @returns an array of bits that represent the number where the first bit is the sign
+ */
+export function packNum(d: number): BitArray {
     const w = decToBin(Math.abs(d), 32)
-
     if (d < 0) {
         w[0] = 1
     }
-
     return w
 }
 
-//Decode literal number
-export function unpackNum(w) {
+/**
+ * Decode literal number (used for the registers)
+ * @param w An array of bits where the first bit is the sign
+ * @returns The bits as a signed decimal number
+ */
+export function unpackNum(w: BitArray): number {
     return (w[0] == 1 ? -1 : 1) * binToDec(w.slice(1))
 }
 
-export function addrToHex(d: number) {
+/**
+ * Converts a decimal number to a 4 character flexidecimal string.
+ * @param d the number to convert in the range of 0-4095
+ * @returns the flexidecimal string representation of the number
+ */
+export function addrToHex(d: number): string {
     const s = String(d).padStart(4, '0')
-
     return halfToHex(decToBin(parseInt(s.substring(0, 2)), 6)) + halfToHex(decToBin(parseInt(s.substring(2, 4)), 6))
 }