+++ ts-lgp-30-html5/src/utils.test.ts
@@ -1,5 +1,5 @@
 import { describe, test, expect } from 'vitest'
-import { insertArrayAt, toLGPHex } from './util'
+import { addrToHex, insertArrayAt, toLGPHex } from './util'
 
 describe("test the utils class", () => {
 
@@ -15,8 +15,20 @@ describe("test the utils class", () => {
     })
 
     test("test insertArrayAt", () => {
-        const arr = insertArrayAt([1, 2, 3, 4, 5], 2, ['a', 'b', 'c'])
-        expect(arr).toStrictEqual([1, 2, 'a', 'b', 'c']) // TODO, insert should extend the array as per below
+        const arr = insertArrayAt([1, 0, 1, 0, 1], 2, [0, 1, 0])
+        expect(arr).toStrictEqual([1, 0, 0, 1, 0]) // TODO, insert should extend the array as per below
         //expect(arr).toStrictEqual([1, 2, 'a', 'b', 'c', 4, 5])
+    })
+
+    test('test addrToHex', () => {
+        expect(addrToHex(0)).toEqual('0000')
+        expect(addrToHex(1)).toEqual('0001')
+        expect(addrToHex(9)).toEqual('0009')
+        expect(addrToHex(10)).toEqual('000f')
+        expect(addrToHex(15)).toEqual('000w')
+        expect(addrToHex(1023)).toEqual('0f17')
+        expect(addrToHex(1024)).toEqual('0f18')
+        // expect(addrToHex(4095)).toEqual('0www') breaks, need to check why
+        // console.debug('4095', addrToHex(4095))
     })
 })